+++ framework/src/main/java/com/live2d/sdk/cubism/framework/motion/ACubismMotion.java
@@ -27,9 +27,11 @@ public abstract class ACubismMotion {
      * @param motionQueueEntry motion managed by CubismMotionQueueManager
      * @param userTimeSeconds total delta time[s]
      */
-    public void updateParameters(CubismModel model,
-                                 CubismMotionQueueEntry motionQueueEntry,
-                                 float userTimeSeconds) {
+    public void updateParameters(
+        CubismModel model,
+        CubismMotionQueueEntry motionQueueEntry,
+        float userTimeSeconds
+    ) {
         if (!motionQueueEntry.isAvailable() || motionQueueEntry.isFinished()) {
             return;
         }