+++ framework/src/main/java/com/live2d/sdk/cubism/framework/rendering/android/CubismClippingContext.java
@@ -25,6 +25,10 @@ class CubismClippingContext {
         final int[] clippingDrawableIndices,
         int clipCount
     ) {
+        if (manager == null || clippingDrawableIndices == null) {
+            throw new IllegalArgumentException("manager or clippingDrawableIndices is null.");
+        }
+
         owner = manager;
 
         // クリップしている（＝マスク用の）Drawableのインデックスリスト
@@ -43,54 +47,6 @@ class CubismClippingContext {
         clippedDrawableIndexList.add(drawableIndex);
     }
 
-    public boolean isUsing() {
-        return isUsing;
-    }
-
-    public void isUsing(boolean isUsing) {
-        this.isUsing = isUsing;
-    }
-
-    public int[] getClippingIdList() {
-        return clippingIdList;
-    }
-
-    public int getClippingIdCount() {
-        return clippingIdCount;
-    }
-
-    public int getLayoutChannelNo() {
-        return layoutChannelNo;
-    }
-
-    public void setLayoutChannelNo(int layoutChannelNo) {
-        this.layoutChannelNo = layoutChannelNo;
-    }
-
-    public CubismRectangle getLayoutBounds() {
-        return layoutBounds;
-    }
-
-    public CubismRectangle getAllClippedDrawRect() {
-        return allClippedDrawRect;
-    }
-
-    public CubismMatrix44 getMatrixForMask() {
-        return matrixForMask;
-    }
-
-    public CubismMatrix44 getMatrixForDraw() {
-        return matrixForDraw;
-    }
-
-    public List<Integer> getClippedDrawableIndexList() {
-        return clippedDrawableIndexList;
-    }
-
-    public CubismClippingManagerAndroid getOwner() {
-        return owner;
-    }
-
     /**
      * このマスクを管理するマネージャのインスタンスを取得する。
      *
@@ -103,39 +59,44 @@ class CubismClippingContext {
     /**
      * クリッピングマスクのIDリスト
      */
-    private final int[] clippingIdList;
+    public final int[] clippingIdList;
     /**
      * 現在の描画状態でマスクの準備が必要ならtrue
      */
-    private boolean isUsing;
+    public boolean isUsing;
     /**
      * クリッピングマスクの数
      */
-    private final int clippingIdCount;
+    public final int clippingIdCount;
     /**
      * RGBAのいずれのチャンネルにこのクリップを配置するか(0:R, 1:G, 2:B, 3:A)
      */
-    private int layoutChannelNo;
+    public int layoutChannelNo;
     /**
      * マスク用チャンネルのどの領域にマスクを入れるか(View座標-1..1, UVは0..1に直す)
      */
-    private final CubismRectangle layoutBounds = CubismRectangle.create();
+    public final CubismRectangle layoutBounds = CubismRectangle.create();
     /**
      * このクリッピングで、クリッピングされる全ての描画オブジェクトの囲み矩形（毎回更新）
      */
-    private final CubismRectangle allClippedDrawRect = CubismRectangle.create();
+    public final CubismRectangle allClippedDrawRect = CubismRectangle.create();
     /**
      * マスクの位置計算結果を保持する行列
      */
-    private final CubismMatrix44 matrixForMask = CubismMatrix44.create();
+    public final CubismMatrix44 matrixForMask = CubismMatrix44.create();
     /**
      * 描画オブジェクトの位置計算結果を保持する行列
      */
-    private final CubismMatrix44 matrixForDraw = CubismMatrix44.create();
+    public final CubismMatrix44 matrixForDraw = CubismMatrix44.create();
     /**
      * このマスクにクリップされる描画オブジェクトのリスト
      */
-    private final List<Integer> clippedDrawableIndexList = new ArrayList<Integer>();
+    public final List<Integer> clippedDrawableIndexList = new ArrayList<Integer>();
+    /**
+     * このマスクが割り当てられるレンダーテクスチャ（フレームバッファ）やカラーバッファのインデックス
+     */
+    public int bufferIndex;
+
     /**
      * このマスクを管理しているマネージャのインスタンス
      */